requirements.txt
Flask==3.0.0
requests==2.31.0
pandas>=2.2.0
numpy>=1.26.0
python-dotenv==1.0.0
scipy>=1.12.0
pytest>=8.0.0
flask-cors==4.0.0
matplotlib>=3.8.0
plotly>=5.17.0
